fixed some documentation, removed debug sysouts, moved contentservice tests from ContentCreation to ContentServiceTest
