+++ src/test/java/com/revature/servicestests/ContentServiceTest.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.Commit;
+import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.jdbc.JdbcTestUtils;
@@ -22,7 +23,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.revature.entities.Content;
 import com.revature.entities.Link;
+import com.revature.entities.Module;
 import com.revature.services.ContentService;
+import com.revature.services.ModuleService;
 import com.revature.services.SearchService;
 
 @ExtendWith(SpringExtension.class)
@@ -40,6 +43,9 @@ class ContentServiceTest {
 
 	@Autowired
 	JdbcTemplate template;
+	
+	@Autowired
+	ModuleService ms;
 
 	int rows = 0;
 
@@ -107,6 +113,42 @@ class ContentServiceTest {
 	@Order(7)
 	void deleteTestData() {
 		JdbcTestUtils.deleteFromTableWhere(template, "content", String.format("title = '%s'", "FIRST TEST CONTENT"));
+	}
+	
+	@Test
+	@Rollback
+	@Order(8)
+	public void updateContent() {
+		
+		Module module1 = new Module(0, "Element Zero", 0, null);
+		module1 = ms.createModule(module1);
+		
+		Module module2 = new Module(0, "Mass Effect Fields", 0, null);
+		module2 = ms.createModule(module2);
+		
+		Set<Link> links = new HashSet<Link>();
+		
+		//we will know the module id but not the content id or link id
+		Link link1 = new Link(0,0,module1.getId(),"relaventTo");
+		Link link2 = new Link(0,0,module2.getId(),"relaventTo");
+		
+		links.add(link1);
+		links.add(link2);
+		
+		Content content = new Content(0,
+				"Mass Effect Theory",
+				"Hologram",
+				"AI explains Mass Effect Theory",
+				"www.mass.test",
+				links);
+		
+		content = cs.createContent(content);
+		
+		content.setDescription("RIP Bioware...");
+		//System.out.println("Getting: " + content.toString());
+		
+		assertNotNull(cs.updateContent(content));
+		assertEquals(0, content.getDescription().compareTo(cs.getContentById(content.getId()).getDescription()));
 	}
 
 }
